requirements.txt
pyfiglet==0.8.post1
python-magic==0.4.27
requests==2.26.0
